[ADD] 일기 커버 작성 페이지를 만들어라 - url에 keyword를 이용해 postPage에서 키워드에 해당하는 커버와 리스트 컴포넌트를 보여주게 하였다 - 네비게이션과 연결된 작성 버튼에도 url을 수정해주었다 - 커버 페이지와, 리스트 페이지를 생성 및 디자인하였다 - (고민)아직 수정페이지는 하지 못했는데, 공통 컴포넌트 설계를 어떻게 할지 고민이다
